+++ app/mailers/claim_mailer.rb
@@ -37,6 +37,14 @@ class ClaimMailer < ApplicationMailer
     send_mail
   end
 
+  def ecp_email_verification(claim, one_time_password)
+    set_common_instance_variables(claim)
+    @subject = "Early-career payment email verification"
+    @one_time_password = one_time_password
+
+    send_mail
+  end
+
   private
 
   def set_common_instance_variables(claim)