+++ spec/models/claim_spec.rb
@@ -807,7 +807,8 @@ RSpec.describe Claim, type: :model do
         :middle_name,
         :surname,
         :banking_name,
-        :building_society_roll_number
+        :building_society_roll_number,
+        :one_time_password
       ])
     end
   end